nixos/modules/sway: Add cap_sys_nice by default
The capability `CAP_SYS_NICE` can be set since Sway 1.8 to improve
scheduling: https://github.com/swaywm/sway/pull/6994

This is optional but we decided to enable it by default as it is useful.
It has some security implications though (but the capabilities are only
for the Sway process itself since we don't make them inheritable).
The definition of `CAP_SYS_NICE` from capabilities(7):
```
* Lower the process nice value (nice(2), setpriority(2)) and change the
  nice value for arbitrary processes;
* set real-time scheduling policies for calling process, and set
  scheduling policies and priorities for arbitrary processes
  (sched_setscheduler(2), sched_setparam(2), sched_setattr(2));
* set CPU affinity for arbitrary processes (sched_setaffinity(2));
* set I/O scheduling class and priority for arbitrary processes
  (ioprio_set(2));
* apply migrate_pages(2) to arbitrary processes and allow processes to
  be migrated to arbitrary nodes;
* apply move_pages(2) to arbitrary processes;
* use the MPOL_MF_MOVE_ALL flag with mbind(2) and move_pages(2).
```

This should help avoiding some stuttering as well as the "your
compositor is too slow" warnings from libinput.