POC: devShell interface
The idea here is that future Nix first tries to "`nix run`" the
`pkg.devShell` package, and only if that fails, fall back to the
legacy `nix develop` (or `nix-shell`) behavior.

This allows the development shell to evolve with stdenv, and it
allows packages to individually customize the `devShell` attribute,
by setting `passthru.devShell`.

Furthermore, these shell behaviors will be pinned to the expressions,
allowing changes to be made in a more agile manner, unlike Nix,
which has to be very careful not to break old expressions, as users
can not revert Nix.

To give it a try:

    nix run .#hello.devShell

In the future this will be equivalent to:

    nix develop .#hello

Isn't this the responsibility of Nix?

It is not. `nix-shell` and `nix develop` are a great user interface,
that everyone loves, but their implementation is a pile of hacks on
top of stdenv.
Instead of coercing stdenv to do what `nix-shell` needs it to, we can
ask stdenv politely to provide a shell.
Now that Nix doesn't have to assume a package comes from stdenv,
there's a possibility for experimental builders to provide shells too.

What does this break?

Only packages that define a `devShell` attribute (for some reason?)
have to adapt to the suggested new Nix behavior. Note that a
`devShell` value for the builder can be overridden by `passthru`
without affecting the build or shell.
Packages and shells pinned to older versions can still be loaded
because Nix keeps the legacy behavior as a fallback.

But this still relies on `nix-shell` to provide a shell???

Fair enough. This is only a proof of concept. The goal is to replace
that invocation by a script or program with the same or better
behavior, without relying on `nix-shell` as its implementation.

Does this solve the need for `.env` for Haskell package shells?

Not in this commit, but Haskell packages will be able to produce
their own `devShell` attribute, which is derived from the .env
derivation rather than the regular derivation.

Refs
 - https://github.com/NixOS/nix/issues/7468 and a bunch of other
   issues where I've preached about this idea.