stdenv: add `developDerivation`
Once NixOS/nix#7480 is merged, this will be an exact hash equivalent
of the derivation which `nix develop` creates directly in C++ code.

Being able to reference this in Nix expressions is essentially to
properly capturing it as a dependency for scripts, and for simply
building the shell as a separate stage before executing it.