git: set reasonable default features
Git should come with a reasonable set of baseline features. Arguable
git-send-email is one of them. Development flows, such as those of
Linux, depend on the tool being available.

The difference in closure size between this and the previous version
is is negligible. There are no additional build-time dependencies
pulled in and the runtime closure difference is about 1%.