ath9k-htc-blobless-firmware: init at 1.4.0
This firmware is completely open source with no blobs, which is
quite rare in the wifi world.  Wifi chips have their own dedicated
general-purpose CPUs.  This source code allows you to see what those
CPUs are doing and modify their behavior.

When the upstream repository was created in 2013, "open source
firmware" meant "firmware which is open source".  In 2023 that is no
longer the generally accepted [definition], so I have chosen an
unambiguous adjective (whose meaning has remained stable for
decades) to use in the pname.

[definition]: https://web.archive.org/web/20221209121315/https://www.opencompute.org/projects/open-system-firmware#:~:text=Another,allows%20it

Co-authored-by: Artturi <Artturin@artturin.com>