olaris-server: 0.4.0 -> unstable-2022-06-11
As called out in the linked issue below, olaris-server has a runtime
dependency on ffmpeg. Unfortunately, 0.4.0 requires a custom fork of
ffmpeg. While we could fetch the upstream compiled artefact or build it
ourselves, the former was unpalatable and the latter prohibitively
difficult. As such, we have bumped to the, yet to be released, tip of
the default branch, which has merged support for upstream ffmpeg.

Fixes #207877