treewide: use nativeBuildInputs with runCommand instead of inlining
This makes cross-compilation not only easier, but possible, as most of these derivations didn't reference the packages correctly to allow the binaries to run on the build systwm.

This is based off 26f704b545838084e334f37d434a648c0c564ffd (#206775).

After this was merged, we noticed the install tests weren't pulling in all the necessary dependencies for a system build. This has happened in the past, but nothing obvious stands out when examining this change.

To summarize: this is because `nativeBuildInputs` always picks the `dev` output of a given derivation. This never ocurred for the other packages, as they either had a `bin` output that was specified, or had no dev output at all.

The exception to this is Nix, which has an `out` output with its binaries, and a `dev` output. So, when using `nativeBuildInputs`, `nix.dev` was referenced instead of `nix.out`, which the test VM's store doesn't contain.

In addition to this fix, I've cleaned up the original commit a bit. Specifically, I:

1. Removed usages of `with pkgs` if it wasn't already in use in the derivation.
2. Removed unrelated formatting changes.
3. Removed an unrelated move from `runCommand` to `runCommandLocal` in `nixos/modules/installer/tools/nixos-build-vms/build-vms.nix`.

Co-authored-by: Sandro <sandro.jaeckel@gmail.com>