emacs: allow overriding version argument
This allows one to use .override to change the version as it's being
used in the derivation builder function, not just as a final derivation
property (which would be .overrideAttrs). The version argument is used
to determine other aspects of the final build, like which patches to
apply: if you pass in a newer source (which you /can/ do using
.overrideAttrs), the derivation builder won't know that and will still
give you the patches for the older source.

Example of what you would want to do:

    pkgs.emacs.overrideAttrs (_: {
      version = "30.0-git";
      src = latest-emacs-src-from-master;
    })

That will override the ‘version’ property on the derivation (good) but
it won't override the ‘version’ argument as used by the rest of the
file, notably the patch selection's if-guard. So you'll end up with a
failed patch.

Example of what you can do after this patch has been applied to nixpkgs:

    pkgs.emacs.override ({
      version = "30.0-git";
    }).overrideAttrs (_: {
      src = latest-emacs-src-from-master;
    })