nixos/pacemaker: updates to allow ocf:heartbeat:IPaddr2 resources
Added/refactored pacemaker tests to test for vip IPaddr2 resource usage. Future
work might be to wrap the config into nixos module options for corosync.

Other ocfs:heartbeat:* resources might need similar work and/or tests added. For
now I only validated IPaddr2 as thats what I need.

I also punted on linking the vip to the cat test resource for now and simply
ping the vip address in the test to see if it is moving or not.