python310Packages.catboost: 1.0.5 -> 1.1.1, use ya.make
Switching to ya allows to do darwin build and fixes the cython issue (which,
despite what the comment says, is not limited to the withCuda=true).

The resulting package doesn't build on darwin with sandbox, as ya-bin tool
contains frozen python2 executable, which requires /usr/sbin/sysctl to detect
number of CPUs in the multiprocessing module.