ventoy-bin: remove `inherit (libsForQt5)` reference on all-packages.nix
According to the issue linked at the end, filling arguments for `callPackage` in
all-packages.nix breaks splicing and makes the expressions incompatible with
cross-compilation.

This issue is more pronounced in the many `inherit (darwin)` calls. However, it
is present in similar contexts too.

The idea here is not the smartest: it just hides those inheritances from
`all-packages.nix` and sends them to the `default.nix` expressions.

Issue: https://github.com/NixOS/nixpkgs/issues/204303