prometheus-xmpp-alerts: Fix broken binary
The binary was broken after the update to 0.5.6 (#205559) due to a
regression in upstream (the module name was changed but not updated in
the appropriate entrypoint). A patch is applied until fixed in upstream.
A very simple test should detect future regressions like this and prevent
breaking the package and associated module.