gnatboot12: Use bintoolsDualAs on x86_64-darwin
As documented in pkgs/os-specific/darwin/binutils/default.nix,
the Clang integrated assembler is needed in addition to the cctools-port
assembler usually used on x86_64-darwin in order to build gnat12 with
the x86_64-darwin gnatboot12 bootstrap compiler.  This commit alters the
gnatboot12 wrapper on x86_64-darwin so that both of those assemblers are
available as required.