stdenv/mkDerivation include dot folders in globs
Related to #208615

unpackFile copies folders into the working directory and the second part of unpackPhase
 then checks whether anything was indeed copied.
It does so by using `for i in *` (a glob).
By default globs in bash do not include items with names that start with a dot ("hidden files").

However, sometimes derivation sources maybe in "hidden" folders and when they are, nix-build fails.

Bash allows their inclusion with the dotglob shell option, which is what this commit does.