chromium: Drop passthru.updateScript
The update script for Chromium does not behave like other update
scripts, e.g., it updates all channels and ungoogled-chromium (not just
a single package). This causes issues when r-ryantm (the bot that runs
nixpkgs-update) tries to update chromium and ungoogled-chromium (which
it started to do recently):
- https://github.com/NixOS/nixpkgs/pull/207185
- https://github.com/NixOS/nixpkgs/pull/207482
- https://github.com/NixOS/nixpkgs/pull/207596
- https://github.com/NixOS/nixpkgs/pull/208048
- https://github.com/NixOS/nixpkgs/pull/208287

We could make the update script work as expected by other tooling but
Chromium requires special attention anyway so it might also be for the
best to do it manually for now (at least someone needs to ensure that
the correct labels are set and that security fixes (~ all updates) are
backported).

(cherry picked from commit c9ca13ce9abb2f872c1a24f49fcc027343146a98)