nss: convert NIX_CFLAGS_COMPILE to a list
Nss has a lot of dependers, so subtle whitespace changes can cause
mass rebuilds, like I did here:

  https://github.com/NixOS/nixpkgs/pull/208698#issuecomment-1368600990

Let's use a list instead of concatenating individual strings in
order to be more robust.  This will cause a lot of rebuilds, but
will prevent rebuilds in the future.  Sending it this cleanup to
staging because of that.