darwin.builder: Instruct user not to disable firewall
It turns out that it's not necessary to disable the firewall
because `qemu` by default binds to all ports (`0.0.0.0`),
including `127.0.0.1`.

Ideally, we would configure `qemu` to bind to `127.0.0.1` so
that the user would not even be prompted to disable their
firewall, but that doesn't work for unprivileged services on
macOS.

You heard that right.  On macOS an unprivileged service can
bind to `0.0.0.0:22` but not `127.0.0.1:22`. See:

https://developer.apple.com/forums/thread/674179

… so right now the lesser evil is just telling users to
not disable the firewall rather than requiring them to run
the script with `sudo`.