Revert "nixos/stage-1: fix `modprobe` in initial ramdisk on systems w/glibc-2.34"
This reverts commit da905d4cf918e28f924bab99c6f2a4012d55410d.

See the commit linked above for further information on why this was
needed. Apparently this is not needed anymore because the need for
LD_LIBRARY_PATH (which is needed for `modprobe(8)` to find
`libpthread.so.0`) doesn't exist anymore.

Since d33e52b2539c5b36a5a876df9006a7145efd42ea the library path of each
binary in extra-utils is patched correctly.