spark2014: fix incomplete install (#208830)
The previous installPhase was only using `make install`, while the real
target needed is `make install-all`. This installs many missing tools
under `$out/libexec/`, such as the missing `gnatwhy3` binary, which was
the root cause of #208830.

With this, gnatprove can now successfully discharge and prove things
with solvers like Z3, etc.

And, because it also includes `why3server`, we don't need to use
wrapProgram on it anymore, either!

Closes #208830

Signed-off-by: Austin Seipp <aseipp@pobox.com>