emscripten: fake cc-wrapper-esque passthru attrs for GHC
For the GHC JavaScript backend, we'll use emscripten in place of
targetCC. To avoid having too much special logic for this, we'll make
the emscripten derivation look like the result of wrapCC as far as GHC
is concerned, i.e. we need targetPrefix and bintools.

For bintools, we'll just reexpose emscripten, as it has emar, the only
relevant bintools. That the other ones are missing doesn't matter in
practice, as the GHC build system won't attempt to use them.

targetPrefix can immediately be (ab)used to make sure GHC will correctly
call emcc etc. instead of plain cc.