stdenv: throw from ghcjs.cc.outputs rather than ghcjs.cc
`pkgsCross.ghcjs.stdenv.cc` currently `throw`s, and it is (a good)
part of the Nix semantics that exceptions transmit no information to
an enclosing `tryEval`.  This makes it impossible to express the
fact that the answer to "is the C compiler from GNU?" should be
`false` when no compiler exists.

The reason why `pkgsCross.ghcjs.stdenv.cc` is a `throw` is explained
here:

  https://github.com/NixOS/nixpkgs/blob/aa371581dd110352b217e214867dd2a83660fa3c/pkgs/stdenv/cross/default.nix#L67

which refers to this check in `splice.nix`:

  https://github.com/NixOS/nixpkgs/blob/aa371581dd110352b217e214867dd2a83660fa3c/pkgs/top-level/splice.nix#L59

Since the result of that check is always passed to `getOutputs`,
which is strict in its argument's `outputs` attribtue, we can simply
move the `throw` deeper into the `stdenv.cc` attrset.  This leaves
room to provide a useful value for attributes like `isGNU` and
`isClang` when the compiler is "none".

Hopefully this removes the need for the change in
b6ac11d8a69ad2a04bbbdbaa287da8ba5f7596d8 from #208947.