lib/types: make `either` expose submodules
With this change, the options in a submodule that is part of an `either`
type are now exposed in the manual.

This implementation should be satisfactory for most use cases, but it
has a bug: if both sides of the `either` are submodules, `getSubModules`
only returns the first type's submodules. Without this, I get "option
already defined" errors coming from `services.tor.settings.DNSPort`,
which has a type of the form `either someSubmodule (listOf someSubmodule)`.
This causes the submodules to be duplicated and I'm not sure how to fix
that right now.