wordpress6_1: init at 6.1.1
release-22.11 currently tracks 6.0 for pkgs.wordpress. This means there
is no way to opt in to the newer version. Similiar to nextcloud and go,
we have broken out a new derivation to opt in. It is a nop for unstable,
and only useful when backported.

Updates #208529

(cherry picked from commit f9d1f80045d3ae01741896127837eba9f1559603)