pynvml: Add patch for finding libnvidia-ml.so.1 on NixOS
This fixes loading the shared library used for interacting with Nvidia
hardware. A similar package, nvidia-ml-py, already provides a patch for
the same thing. This commit adds a similar patch for pynvml. We could
not reuse the existing patch for the other package because the files
being patched have different names despite the patch itself being nearly
the same.

Fixes https://github.com/NixOS/nixpkgs/issues/209416