katago: 1.11.0 -> 1.12.3
Added TensorRT backend, refactored backend selection, IMPORTANT also
bumps TensorRT version to 8.5.2 for CUDA 11.7 and 11.8.  (Per Nvidia's
documentation, 8.4.0.6 does not officially support 11.7, but we're using
for it.  Additionally, KataGo requires 8.5 or better for the new
TensorRT backend.  8.4.0.6 remains the default for 11.6 and lower, in
interest of not messing things up as much as possible.)

New version also adds support for nested residual block networks, such as the
b18 network that was used in the UEC.