udev: set to libudev-zero on static linux
systemd does not support being built statically[1], which means that,
prior to this patch, nothing that depends on udev can be built
statically.  libudev-zero is a daemonless implementation of libudev,
that supports being built statically.  So, by setting udev to be
libudev-zero when systemd is unavailable, we can get static builds of
packages like linuxPackages.usbip.

[1]: https://github.com/systemd/systemd/issues/20600#issuecomment-912338965